Add command line flag to select the type of the scalar. Updated several locations to allow for integer scalar types without compile errors. Note that this work is still incomplete; integer scalar types still have many issues and are generally untested.

+++ calculator/parser.h
@@ -42,9 +42,9 @@ struct ExpressionAst : Ast {
 };
 
 struct ScalarAst final : ExpressionAst {
-  float value;
+  double value;
 
-  explicit ScalarAst(float value) : value(value) {}
+  explicit ScalarAst(double value) : value(value) {}
 
   void visit(Visitor& v) override;
 };
